Replace `half2` with `float2` in shaders

+++ gamedata/shaders/r2/shadow.h
@@ -27,7 +27,7 @@ half 	sample_sw	(float2 tc, float2 shift, float depth_cmp)
 		depth_cmp-tex2D	(s_smap, Tex11).x);
 	float4 	compare = step	(depth,0);
 	float2 	fr 		= frac	(Tex00*texsize);
-	half2 	ifr 	= half2	(1,1) - fr;
+	float2 	ifr 	= float2	(1,1) - fr;
 	float4 	fr4 	= float4	(ifr.x*ifr.y, ifr.x*fr.y, fr.x*ifr.y,  fr.x*fr.y);
 	return	dot		(compare, fr4);
 }
@@ -87,7 +87,7 @@ half  	sample_hw_f4	(float4 tc,float4 shift){
 
 	float  	texsize 		= 	SMAP_size	;
 	float2 	fr 				= 	frac		(T4.xy * texsize);
-	half2 	ifr 			= 	half2		(1,1) - fr;
+	float2 	ifr 			= 	float2		(1,1) - fr;
 	float4 	fr4 			= 	float4		(ifr.x*ifr.y, ifr.x*fr.y, fr.x*ifr.y,  fr.x*fr.y);
 	float4 	fr4s		 	= 	fr4.zywx	;
 
@@ -114,7 +114,7 @@ uniform sampler2D	jitter1;
 uniform sampler2D	jitter2;
 uniform sampler2D	jitter3;
 uniform float4 		jitterS;
-float4 	test 		(float4 tc, half2 offset)
+float4 	test 		(float4 tc, float2 offset)
 {
 	float4	tcx	= float4 (tc.xy + tc.w*offset, tc.zw);
 	return 	tex2Dproj (s_smap,tcx);
@@ -147,10 +147,10 @@ half 	shadowtest_sun 	(float4 tc, float4 tcJ)			// jittered sampling
 	//float4	J1 	= tex2D	(jitter1,tc_J)*scale;
 
 	const float k = .5f/float(SMAP_size);
-	r.x 	= test 	(tc, J0.xy+half2(-k,-k)).x;
-	r.y 	= test 	(tc, J0.wz+half2( k,-k)).y;
-	r.z		= test	(tc,-J0.xy+half2(-k, k)).z;
-	r.w		= test	(tc,-J0.wz+half2( k, k)).x;
+	r.x 	= test 	(tc, J0.xy+float2(-k,-k)).x;
+	r.y 	= test 	(tc, J0.wz+float2( k,-k)).y;
+	r.z		= test	(tc,-J0.xy+float2(-k, k)).z;
+	r.w		= test	(tc,-J0.wz+float2( k, k)).x;
 
 	return	dot(r,1.h/4.h);
 }
